+++ src/main/java/bridge/BridgeGame.java
@@ -1,10 +1,11 @@
-package bridge.service;
+package bridge;
 
 import bridge.domain.GameData;
 import bridge.dto.HistoryMapDto;
 import bridge.dto.PlayerLocationDto;
 import bridge.dto.PlayerTryCountDto;
 import bridge.exception.InputException;
+import bridge.service.PlayerCommend;
 
 /**
  * 다리 건너기 게임을 관리하는 클래스
