+++ src/main/java/bridge/BridgeMaker.java
@@ -1,4 +1,6 @@
-package bridge.service;
+package bridge;
+
+import bridge.service.BridgeNumberGenerator;
 
 import java.util.ArrayList;
 import java.util.Collections;
